Cumulative Hazard Function

Access a distribution's cumulative hazard function (chf).

Usage

eval_chf(distribution, at)

enframe_chf(..., at, arg_name = ".arg", fn_prefix = "chf", sep = "_")

Arguments

distribution, ...

A distribution, or possibly multiple distributions in the case of ....

at

Vector of values to evaluate the representation at.

arg_name

For enframe_, name of the column containing the function arguments. Length 1 character vector.

fn_prefix

For enframe_, name of the function to appear in the column(s). Length 1 character vector.

sep

When enframe'ing more than one distribution, the character that will be separating the fn_name and the distribution name. Length 1 character vector.

Value

The evaluated representation in vector form (for eval_) with length matching the length of at, and data frame or tibble form (for enframe_) with number of rows matching the length of at. The at input occupies the first column, named .arg by default, or the specification in arg_name; the evaluated representations for each distribution in ... go in the subsequent columns (one column per distribution). For a single distribution, this column is named according to the representation by default (cdf, survival, quantile, etc.), or the value in fn_prefix. For multiple distributions, unnamed distributions are auto-named, and columns are named <fn_prefix><sep><distribution_name> (e.g., cdf_distribution1).

See also

Examples

d <- dst_unif(0, 4)
eval_chf(d, at = 0:4)
#> [1] 0.0000000 0.2876821 0.6931472 1.3862944       Inf
enframe_chf(d, at = 0:4)
#> # A tibble: 5 × 2
#>    .arg     chf
#>   <int>   <dbl>
#> 1     0   0    
#> 2     1   0.288
#> 3     2   0.693
#> 4     3   1.39 
#> 5     4 Inf