Specific Formulas for Quantities

Formulas for quantities (such as mean, variance, skewness, EVI, etc.) of select parametric distributions.

.quantities

Format

An object of class list of length 18.

Details

A list, where each distribution gets a (named) entry, with name given by the suffix of dst_ (such as "norm", "unif", etc.). Each distribution's entry is itself a named list of expressions, where the name is the name of the quantity matching the distionary function name:

mean

median

variance

skewness

kurtosis_exc

range

evi

Each expression is allowed to refer to the distribution's parameters by name.

Note

Although R allows us to evaluate distributional representations of certain parametric distributions through functions with p, d, q, and r prefixes (such as pnorm(), dnorm(), etc.), R does not "come with" formulas for quantities such as mean, variance, EVI, etc. Although these quantities can be computed from a distributional representation (such as integrating the quantile function to get the mean), it's often inefficient to rely on such computations. We therefore include formulas here, and check them using testthat.