Hazard Function

Access a distribution's hazard function.

Usage,

eval_hazard(distribution, at)

enframe_hazard(..., at, arg_name = ".arg", fn_prefix = "hazard", sep = "_")

Arguments

distribution

A distribution, or possibly multiple distributions in the case of ....

at

Vector of values to evaluate the cdf at. Must be named when using in enframe_.

...

A distribution, or possibly multiple distributions in the case of ....

arg_name

For enframe_, name of the column containing the function arguments.

fn_prefix

For enframe_, name of the function to appear in the column(s).

sep

When enframe'ing more than one distribution, the character that will be separating the fn_name and the distribution name.

Value

The evaluated hazard in vector form (for eval_) and data frame or tibble form (for enframe_).

See also

Examples

d <- dst_unif(0, 4)
eval_hazard(d, at = 0:4)
#> [1] 0.2500000 0.3333333 0.5000000 1.0000000       Inf
enframe_hazard(d, at = 0:4)
#> # A tibble: 5 × 2
#>    .arg  hazard
#>   <int>   <dbl>
#> 1     0   0.25 
#> 2     1   0.333
#> 3     2   0.5  
#> 4     3   1    
#> 5     4 Inf