Return Level Function

Compute return levels (quantiles) from a distribution by inputting return periods.

Usage

eval_return(distribution, at)

enframe_return(..., at, arg_name = ".arg", fn_prefix = "return", sep = "_")

Arguments

distribution, ...

A distribution, or possibly multiple distributions in the case of ....

at

Vector of return periods >=1.

arg_name

For enframe_, name of the column containing the function arguments.

fn_prefix

For enframe_, name of the function to appear in the column(s).

sep

When enframe'ing more than one distribution, the character that will be separating the fn_name and the distribution name.

Value

The return levels associated with the specified return periods in vector form (for eval_) and data frame or tibble form (for enframe_). This function is simply the quantile function evaluated at 1 - 1 / at.

See also

Examples

d <- dst_unif(0, 4)
eval_return(d, at = c(2, 25, 100, 200))
#> [1] 2.00 3.84 3.96 3.98
enframe_chf(d, at = c(2, 25, 100, 200))
#> # A tibble: 4 × 2
#>    .arg     chf
#>   <dbl>   <dbl>
#> 1     2   0.693
#> 2    25 Inf    
#> 3   100 Inf    
#> 4   200 Inf